RAG/requirements.txt
langchain
langchain_core
langchain_openai
python-dotenv
pypdf
faiss-cpu
langchainhub
langchain_community
flask
flask-cors
langgraph
Flask[async]